# docs/openapi/swaggerhub-export.json
{
  "paths": {
    "/coredata": {
      "get": {
        "tags": [
          "visitor"
        ],
        "summary": "Gets core data for site",
        "operationId": "coredata",
        "description": "Gets core data for front - available countries, currencies and langs and their default values\n",
        "responses": {
          "200": {
            "description": "search results matching criteria",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Coredata"
                },
                "example": {
                  "value": {
                    "lang": {
                      "code": "en",
                      "longCode": "en-US",
                      "name": "English"
                    },
                    "langs": [
                      {
                        "code": "sk",
                        "longCode": "sk-SK",
                        "name": "Slovenčina"
                      },
                      {
                        "code": "en",
                        "longCode": "en-US",
                        "name": "English"
                      }
                    ],
                    "country": {
                      "code": "sk",
                      "name": "Slovakia"
                    },
                    "countries": [
                      {
                        "code": "sk",
                        "name": "Slovakia"
                      },
                      {
                        "code": "us",
                        "name": "USA"
                      }
                    ],
                    "currency": {
                      "code": "EUR",
                      "symbol": "€"
                    },
                    "currencies": [
                      {
                        "code": "EUR",
                        "symbol": "€"
                      },
                      {
                        "code": "USD",
                        "symbol": "$"
                      }
                    ],
                    "translation": {
                      "en": [
                        {
                          "original": "Price"
                        },
                        {
                          "selector": "example: html > body > template > article.tile.is-child.box > p:nth-child(5) > span"
                        },
                        {
                          "string": "Cena"
                        },
                        {
                          "type": "text"
                        }
                      ]
                    },
                    "user": {
                      "id": "123456789abcdef",
                      "username": "User",
                      "email": "email@example.com",
                      "type": "user",
                      "token": "1234567890abcdefghijklmnopqrstuvwxyz$1234567890abcdefghijklmnopqrstuvwxyz$",
                      "image": "static/users/user.jpg",
                      "bio": "I am programmer",
                      "addresses": [],
                      "settings": {
                        "currency": "EUR",
                        "language": "sk"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "out of service"
          }
        }
      }
    },
    "/coredata/translation": {
      "get": {
        "tags": [
          "visitor"
        ],
        "summary": "Gets translation strings",
        "operationId": "coredatatranslation",
        "description": "Gets translation strings for selected language\n",
        "parameters": [
          {
            "in": "query",
            "name": "lang",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "Language code eg. \"sk\" or \"en\""
          },
          {
            "in": "query",
            "name": "blockName",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "Name of the translation block"
          }
        ],
        "responses": {
          "200": {
            "description": "search results matching criteria",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Translation"
                    }
                  }
                },
                "example": {
                  "value": {
                    "sk": [
                      {
                        "original": "Price",
                        "selector": "html > body > template > article.tile.is-child.box > p:nth-child(5) > span",
                        "string": "Cena",
                        "type": "text"
                      }
                    ]
                  }
                }
              }
            }
          },
          "404": {
            "description": "out of service"
          }
        }
      },
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Update translation strings",
        "operationId": "coredatatranslationupdate",
        "description": "Updates translation strings\n",
        "parameters": [
          {
            "in": "query",
            "name": "dictionary",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Json object of"
          }
        ],
        "responses": {
          "200": {
            "description": "Information about success or error of saving dictionary",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "value": {
                    "success": true
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "tags": [
          "visitor"
        ],
        "summary": "Login user",
        "operationId": "userLogin",
        "description": "Send user credentials to login\n",
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "password",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "requestBody": {
          "description": "username and password",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLoginRequest"
              },
              "example": {
                "user": {
                  "email": "email@example.com",
                  "password": "userPass"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "search results matching criteria",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserScheme"
                },
                "example": {
                  "user": {
                    "_id": "123456789abcdef",
                    "username": "User",
                    "email": "email@example.com",
                    "type": "user",
                    "bio": "I am programmer",
                    "image": "/static/users/user.jpg",
                    "activated": "2019-01-31T18:24:36.274Z",
                    "addresses": [
                      {
                        "type": "invoice",
                        "nameFirst": "User",
                        "nameLast": "Tester",
                        "street": "Street 1",
                        "street2": "",
                        "zip": "98765",
                        "city": "City of Town",
                        "country": "UK",
                        "phone": "987654321",
                        "state": ""
                      }
                    ],
                    "token": "1234567890abcdefghijklmnopqrstuvwxyz$1234567890abcdefghijklmnopqrstuvwxyz$"
                  }
                }
              }
            }
          },
          "400": {
            "description": "out of service"
          }
        }
      }
    },
...
    "/users/checkemail": {
      "post": {
        "tags": [
          "visitor"
        ],
        "summary": "Email check",
        "operationId": "checkemail",
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "requestBody": {
          "description": "check if email is registered",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "name@domain.tld"
                  }
                }
              },
              "example": {
                "email": "name@domain.tld"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "email not foud",
            "content": {
              "application/json": {
                "example": {
                  "result": {
                    "emailExists": false
                  }
                }
              }
            }
          },
          "422": {
            "description": "email exists",
            "content": {
              "application/json": {
                "example": {
                  "errors": {
                    "email": "exists"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/checkusername": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "User name check",
        "operationId": "checkUsername",
        "parameters": [
          {
            "in": "query",
            "name": "username",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "requestBody": {
          "description": "check if username is used",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username"
                ],
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "someusername"
                  }
                }
              },
              "example": {
                "username": "someusername"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "username not foud",
            "content": {
              "application/json": {
                "example": {
                  "result": {
                    "userExists": false
                  }
                }
              }
            }
          },
          "422": {
            "description": "username exists",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "errors": {
                    "username": "exists"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/register": {
      "post": {
        "tags": [
          "visitor"
        ],
        "summary": "User registration",
        "operationId": "userRegister",
        "parameters": [
          {
            "in": "query",
            "name": "username",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "email",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "password",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "settings",
            "required": true,
            "example": {
              "language": "sk",
              "currency": "EUR"
            },
            "schema": {
              "type": "object",
              "properties": {
                "language": {
                  "type": "string"
                },
                "currency": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "requestBody": {
          "description": "Send user registration",
          "required": true,
          "content": {
            "application/data": {
              "schema": {
                "$ref": "#/components/schemas/UserRegistration"
              },
              "example": {
                "username": "MyName",
                "email": "user@domain.tld",
                "password": "PasSW0rD",
                "settings": {
                  "language": "sk",
                  "currency": "EUR"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "user registered",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "user": {
                    "_id": "1234567890abcdef12345678",
                    "username": "MyName",
                    "email": "user@domain.tld",
                    "type": "user",
                    "bio": "",
                    "image": "",
                    "settings": {
                      "language": "sk",
                      "currency": "EUR"
                    },
                    "token": "abcdef1234567890"
                  }
                }
              }
            }
          },
          "422": {
            "description": "username exists",
            "content": {
              "application/json": {
                "example": {
                  "oneOf": [
                    {
                      "errors": {
                        "username": "is exist"
                      }
                    },
                    {
                      "errors": {
                        "email": "is exist"
                      }
                    },
                    {
                      "errors": {
                        "fieldName": "The 'fieldName' field is required!"
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/user": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get profile of logged user",
        "operationId": "userMe",
        "responses": {
          "200": {
            "description": "Profile of logged user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserScheme"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "user"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "user",
            "schema": {
              "$ref": "#/components/schemas/UserScheme"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Create new user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserScheme"
                }
              }
            }
          },
          "422": {
            "description": "User not updated - missing fields or not authorized for that operation"
          }
        }
      }
    },
    "/user/verify": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Verify user email",
        "operationId": "userVerify",
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "hash",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "action",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Email verified",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserScheme"
                }
              }
            }
          },
          "422": {
            "description": "Activation failed"
          }
        }
      }
    },
    "/user/reset": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Reset user password",
        "operationId": "userReset",
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Password reset done, email sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserScheme"
                }
              }
            }
          },
          "422": {
            "description": "Account reset failed"
          }
        }
      }
    },
    "/user/image": {
      "post": {
        "tags": [
          "user"
        ],
        "operationId": "userImageUpload",
        "parameters": [
          {
            "in": "query",
            "name": "file",
            "schema": {
              "type": "string",
              "format": "binary"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Upload user profile image",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "files[0]": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/image/{type}/{code}/{image}": {
      "delete": {
        "tags": [
          "user"
        ],
        "operationId": "userImageDelete",
        "parameters": [
          {
            "in": "path",
            "name": "type",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "code",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "image",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Delete image uploaded by user. Admin can delete all images, other users can delete only their images.",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "message": "file deleted"
                }
              }
            }
          },
          "422": {
            "description": "Couldn't delete file",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "delete failed"
                }
              }
            }
          }
        }
      }
    },
...
    "/cart": {
      "get": {
        "tags": [
          "visitor"
        ],
        "summary": "Get cart to selected visitor",
        "operationId": "cartGet",
        "responses": {
          "200": {
            "description": "Cart content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "visitor"
        ],
        "summary": "Update amount of product in cart",
        "operationId": "cartPostAmount",
        "parameters": [
          {
            "in": "query",
            "name": "itemId",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "amount",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "requestBody": {
          "description": "Send product id and amount",
          "required": true,
          "content": {
            "application/data": {
              "schema": {
                "$ref": "#/components/schemas/CartUpdate"
              },
              "example": {
                "itemId": "abcdef1234567890",
                "amount": 1
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product amount changed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "visitor"
        ],
        "summary": "Add product in the cart",
        "operationId": "cartPutProduct",
        "parameters": [
          {
            "in": "query",
            "name": "itemId",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "amount",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "requestBody": {
          "description": "Send product id and amount",
          "required": true,
          "content": {
            "application/data": {
              "schema": {
                "$ref": "#/components/schemas/CartUpdate"
              },
              "example": {
                "itemId": "abcdef1234567890",
                "amount": 1
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product added to cart",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "visitor"
        ],
        "summary": "Remove the whole cart",
        "operationId": "cartRemoveCart",
        "responses": {
          "200": {
            "description": "Empty cart",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          }
        }
      }
    },
    "/cart/{itemId}": {
      "delete": {
        "tags": [
          "visitor"
        ],
        "summary": "Remove the product from cart",
        "operationId": "cartRemoveProduct",
        "parameters": [
          {
            "in": "path",
            "name": "itemId",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Product removed from cart",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          }
        }
      }
    },
    "/cart/{itemId}/{amount}": {
      "delete": {
        "tags": [
          "visitor"
        ],
        "summary": "Remove the product amount from cart",
        "operationId": "cartRemoveProductAmount",
        "parameters": [
          {
            "in": "path",
            "name": "itemId",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "amount",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Product amount removed from cart",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "openapi": "3.0.0",
  "info": {
    "description": "Main API for StretchShop application",
    "version": "1.0.0-oas3",
    "title": "StretchShop main API",
    "contact": {
      "email": "support@stretchshop.app"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  }
}